+++ AAEmu.Game/Core/Packets/G2C/SCTrialCancledPacket.cs
@@ -3,7 +3,7 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
-public class SCTrialCancledPacket(uint trialId) : GamePacket(SCOffsets.SCTrialCancledPacket, 1)
+public class SCTrialCancledPacket(ulong trialId) : GamePacket(SCOffsets.SCTrialCancledPacket, 1)
 {
     public override PacketStream Write(PacketStream stream)
     {
